The changes all relate to updating the add_to_email_list(), to the point of making it superfluous. The add_to_email_list() method is similar to the much newer method, add() method in D::ML::Subscribers, except, it supports adding more than one email address at a time, has a bunch of options not used, nor documented in the code and also does NOT support subscriber fields.
That last point is a killer - and in the codebase, this method is only used out of shear laziness, or to work with the black_list, which, for some reason, wasn't supported in the, D::ML::Subscribers add_email method (which is not the D::ML::Subscriber add() method) 

So, it's useless - but I'll keep it around for a while for backwards compat. 

Rewriting this one method has removed hundreds of lines of sourcecode, including multiple implementations of the same method for all FOUR supported backends (kill me) and finally killed some ugly nasty bug-ridden code I never wanted to touch, because of its ugly, nasty bug-ridden nature. 

There may be a small performance penalty, as, I guess...? there was some optimizations with allowing you to pass more than one email address at a time, but it's much much too early to do optimizations for these types of things - cleanup first. 

The next step is remove the lingering uses of the add_to_email_list() method and make sure the tests pass.