Moved many of the new files around to get a better (meaning, simpler) API. Something along the lines of:
DADA
-> Profile
-> -> Fields

and, 

DADA::MailingList
-> Subscribers
-> Subscirber

Since Dada now has, "Profiles" and, "Mailing Lists". "Mailing Lists" have, "Subscribers" 

The relationship between a profile and a subscriber is sort of odd - a subscriber is something a mailing list can have, but a profile can have 0 or more *subscriptions*

A subscription is a subscriber who belongs to a mailing list. 

Although imperfect (maybe?) it's a lot better than saying a DADA::MailingList::Subscriber has a profile and fields that are shared between multiple mailing lists.... 

I think?