More work documenting and writing tests for DADA::Profile::Fields.
A little confused myself about what constitutes a Profile, what a Profile Field (or profile fieldS) are and how attributes fit in all this. THAT all needs explaining in the dos, since they're confused as well. 

(most) all tests should still be passing. A few new ones for the new methods that are documented and all that.