Added an undocumented CL interface. I think it sucks, but it may just work...
I think my plan is to documented it and still call it, "undocumented" so not have to support it - it's not out of, "shitty first draft" mode and I don't have the patience to take it out of that. :/ Using CGI's param() to pass variables between subroutines is particularily offensive. Aww, well.