This changes the behavior of the .htaccess file creation:
Now, if an .htaccess file is already found, it'll be looked at for the directives specific for this plugin to work, and if there are none, will add it's own. If it does find its own directives, it'll change the .htaccess, for only those directives. 

Before, the entire .htaccess file would be rewritten.