Work for #426 - should help fix problems with encoding when sending messages in the, "Send a Message" screen, and "Send an Invitation" screen.
No changes in behavior will be seen in the, "Send a Webpage" screen, as messages are constructed using MIME::Lite::HTML. I think the underlying problem is with using MIME::Lite - it doesn't look like it's marked as being worked on, and the current maintainer notes getting weird bugs that they're not going to fix, but states alternatives. One of them, MIME::Entity, is the one we're using, as it's already used in other places in the app.