This adds URL validation on the server side to all the URLs in the Mailing List: List Information screen.
Want to add similar stuff to the Send a Message screen(s), but I do not want, under any circumstances the problem where a draft isn't saved if the URL is not correct - that seems... unwise. I'll have to double-check that the URL is validated before a mass mailing message is created though (I think it is). 

This DOES make me feel as if sending that email notification for scheduled mailings if they work/fail is a real good idea.